+++ Client/CSprite.h
@@ -12,14 +12,14 @@ class Sprite
 		Sprite(float x, float y);
 		~Sprite();
 
-		bool Init(LPDIRECT3DDEVICE9 device, std::string filename, int width, int height);
-		bool Init(LPDIRECT3DDEVICE9 device, wint_t resource, int width, int height);
+		bool Init(LPDIRECT3DDEVICE9 device, std::string filename, float width, float height);
+		bool Init(LPDIRECT3DDEVICE9 device, wint_t resource, float width, float height);
 		bool begin(LPDIRECT3DDEVICE9 device);
 		bool isInitialized();
 
 		void Draw();
-		void Resize(int width, int height);
-		void Move(int x, int y);
+		void Resize(float width, float height);
+		void Move(float x, float y);
 		void Rotate(float angle);
 
 		void Reset();
